move get_available_models to model exec request to avoid circular imports.
