+++ python/lib/communication/dmod/communication/maas_request/__init__.py
@@ -1,5 +1,4 @@
 from .utilities import (
-    get_available_models,
     get_available_outputs,
     get_distribution_types,
     get_parameters,
@@ -9,7 +8,7 @@ from .distribution import Distribution
 from .parameter import Scalar
 from .external_request import ExternalRequest
 from .external_request_response import ExternalRequestResponse
-from .model_exec_request import ModelExecRequest
+from .model_exec_request import ModelExecRequest, get_available_models
 from .model_exec_request_response import ModelExecRequestResponse
 from .nwm import NWMRequest, NWMRequestResponse
 from .ngen import NGENRequest, NGENRequestResponse
